feat: Live Text (Low performance)
