+++ .github/workflows/deploy.yml
@@ -6,7 +6,7 @@ on:
     types: [closed]
 env:
   DEVELOPER_DIR: /Applications/Xcode_13.2.1.app
-  APP_VERSION: '2.1.3'
+  APP_VERSION: '2.2.0'
   SCHEME_NAME: 'EhPanda'
   ALTSTORE_JSON_PATH: './AltStore.json'
   BUILDS_PATH: '/tmp/action-builds'
