+++ EhPanda/View/Reading/ReadingStore.swift
@@ -14,13 +14,11 @@ struct ReadingState: Equatable {
         case hud
         case share(UIImage)
         case readingSetting
-        case textRecognition(UIImage)
     }
     enum ImageAction {
         case copy
         case save
         case share
-        case textRecognition
     }
     struct CancelID: Hashable {
         let id = String(describing: ReadingState.CancelID.self)
@@ -28,6 +26,7 @@ struct ReadingState: Equatable {
 
     @BindableState var route: Route?
     let gallery: Gallery
+    var galleryDetail: GalleryDetail?
 
     var readingProgress: Int = .zero
     var forceRefreshID: UUID = .init()
@@ -120,7 +119,6 @@ enum ReadingAction: BindableAction {
     case saveImage(URL)
     case saveImageDone(Bool)
     case shareImage(URL)
-    case recognizeTextFromImage(URL)
     case fetchImage(ReadingState.ImageAction, URL)
     case fetchImageDone(ReadingState.ImageAction, Result<UIImage, Error>)
 
@@ -251,9 +249,6 @@ let readingReducer = Reducer<ReadingState, ReadingAction, ReadingEnvironment> { 
     case .shareImage(let imageURL):
         return .init(value: .fetchImage(.share, imageURL))
 
-    case .recognizeTextFromImage(let imageURL):
-        return .init(value: .fetchImage(.textRecognition, imageURL))
-
     case .fetchImage(let action, let imageURL):
         return environment.imageClient.fetchImage(url: imageURL)
             .map({ ReadingAction.fetchImageDone(action, $0) })
@@ -273,9 +268,6 @@ let readingReducer = Reducer<ReadingState, ReadingAction, ReadingEnvironment> { 
                     .saveImageToPhotoLibrary(image).map(ReadingAction.saveImageDone)
             case .share:
                 return .init(value: .setNavigation(.share(image)))
-
-            case .textRecognition:
-                return .init(value: .setNavigation(.textRecognition(image)))
             }
         } else {
             state.hudConfig = .error
@@ -309,6 +301,7 @@ let readingReducer = Reducer<ReadingState, ReadingAction, ReadingEnvironment> { 
         return .merge(effects)
 
     case .fetchDatabaseInfos:
+        state.galleryDetail = environment.databaseClient.fetchGalleryDetail(gid: state.gallery.id)
         return environment.databaseClient.fetchGalleryState(gid: state.gallery.id)
             .map(ReadingAction.fetchDatabaseInfosDone).cancellable(id: ReadingState.CancelID())
 
@@ -561,11 +554,6 @@ let readingReducer = Reducer<ReadingState, ReadingAction, ReadingEnvironment> { 
         return .none
     }
 }
-.haptics(
-    unwrapping: \.route,
-    case: /ReadingState.Route.textRecognition,
-    hapticClient: \.hapticClient
-)
 .haptics(
     unwrapping: \.route,
     case: /ReadingState.Route.readingSetting,
