+++ EhPanda/View/Reading/ReadingView.swift
@@ -18,6 +18,7 @@ struct ReadingView: View {
     @Binding private var setting: Setting
     private let blurRadius: Double
 
+    @StateObject private var liveTextHandler = LiveTextHandler()
     @StateObject private var autoPlayHandler = AutoPlayHandler()
     @StateObject private var gestureHandler = GestureHandler()
     @StateObject private var pageHandler = PageHandler()
@@ -67,6 +68,7 @@ struct ReadingView: View {
                 showsPanel: viewStore.binding(\.$showsPanel),
                 showsSliderPreview: viewStore.binding(\.$showsSliderPreview),
                 sliderValue: $pageHandler.sliderValue, setting: $setting,
+                enablesLiveText: $liveTextHandler.enablesLiveText,
                 autoPlayPolicy: .init(get: { autoPlayHandler.policy }, set: setAutoPlayPolocy),
                 range: 1...Float(viewStore.gallery.pageCount),
                 previewURLs: viewStore.previewURLs,
@@ -103,10 +105,6 @@ struct ReadingView: View {
             .accentColor(setting.accentColor).tint(setting.accentColor)
             .autoBlur(radius: blurRadius).navigationViewStyle(.stack)
         }
-        .sheet(unwrapping: viewStore.binding(\.$route), case: /ReadingState.Route.textRecognition) { route in
-            TextRecognitionView(image: route.wrappedValue)
-                .accentColor(setting.accentColor).autoBlur(radius: blurRadius)
-        }
         .sheet(unwrapping: viewStore.binding(\.$route), case: /ReadingState.Route.share) { route in
             ActivityView(activityItems: [route.wrappedValue])
                 .accentColor(setting.accentColor).autoBlur(radius: blurRadius)
@@ -137,6 +135,9 @@ struct ReadingView: View {
                 page.update(.new(index: newValue))
                 Logger.info("Pager.update", context: ["update": newValue])
             }
+            if liveTextHandler.enablesLiveText {
+                analyzeImageForLiveText(index: .init(sliderValue))
+            }
         }
         .onChange(of: viewStore.readingProgress) { readingProgress in
             Logger.info("viewStore.readingProgress changed", context: ["readingProgress": readingProgress])
@@ -155,6 +156,16 @@ struct ReadingView: View {
             setAutoPlayPolocy(.off)
         }
 
+        // LiveText
+        .onChange(of: liveTextHandler.enablesLiveText) { newValue in
+            Logger.info("liveTextHandler.enablesLiveText changed", context: ["newValue": newValue])
+            if newValue {
+                analyzeImageForLiveText(index: .init(pageHandler.sliderValue))
+            } else {
+                liveTextHandler.cancelRequests()
+            }
+        }
+
         // Orientation
         .onChange(of: setting.enablesLandscape) { newValue in
             Logger.info("setting.enablesLandscape changed", context: ["newValue": newValue])
@@ -165,7 +176,10 @@ struct ReadingView: View {
         .animation(.default, value: gestureHandler.scale)
         .animation(.default, value: viewStore.showsPanel)
         .statusBar(hidden: !viewStore.showsPanel)
-        .onDisappear { setAutoPlayPolocy(.off) }
+        .onDisappear {
+            liveTextHandler.cancelRequests()
+            setAutoPlayPolocy(.off)
+        }
         .onAppear { viewStore.send(.onAppear(setting.enablesLandscape)) }
     }
 
@@ -176,6 +190,7 @@ struct ReadingView: View {
             index: index, isDualPage: isDualPage, isDatabaseLoading: viewStore.databaseLoadingState != .idle,
             backgroundColor: backgroundColor, config: imageStackConfig, imageURLs: viewStore.imageURLs,
             originalImageURLs: viewStore.originalImageURLs, loadingStates: viewStore.imageURLLoadingStates,
+            enablesLiveText: liveTextHandler.enablesLiveText, liveTextGroups: liveTextHandler.liveTextGroups,
             fetchAction: { viewStore.send(.fetchImageURLs($0)) },
             refetchAction: { viewStore.send(.refetchImageURLs($0)) },
             prefetchAction: { viewStore.send(.prefetchImages($0, setting.prefetchLimit)) },
@@ -184,19 +199,45 @@ struct ReadingView: View {
             loadFailedAction: { viewStore.send(.onWebImageFailed($0)) },
             copyImageAction: { viewStore.send(.copyImage($0)) },
             saveImageAction: { viewStore.send(.saveImage($0)) },
-            shareImageAction: { viewStore.send(.shareImage($0)) },
-            recognizeTextFromImageAction: { viewStore.send(.recognizeTextFromImage($0)) }
+            shareImageAction: { viewStore.send(.shareImage($0)) }
         )
     }
 }
 
-// MARK: AutoPlay
+// MARK: Handler methods
 extension ReadingView {
     func setAutoPlayPolocy(_ policy: AutoPlayPolicy) {
         autoPlayHandler.setPolicy(policy, updatePageAction: {
             page.update(.next)
             Logger.info("Pager.update", context: ["update": "next"])
         })
+    }
+    func analyzeImageForLiveText(index: Int) {
+        Logger.info("analyzeImageForLiveText", context: ["index": index])
+        let liveTextGroups = liveTextHandler.liveTextGroups[index]
+        guard liveTextGroups == nil || liveTextGroups?.isEmpty == true else {
+            Logger.info("analyzeImageForLiveText duplicated", context: ["index": index])
+            return
+        }
+        guard let key = viewStore.imageURLs[index]?.absoluteString else {
+            Logger.info("analyzeImageForLiveText URL not found", context: ["index": index])
+            return
+        }
+        KingfisherManager.shared.cache.retrieveImage(forKey: key) { result in
+            switch result {
+            case .success(let result):
+                if let image = result.image?.cgImage {
+                    liveTextHandler.analyzeImage(
+                        image, index: index, recognitionLanguages:
+                            viewStore.galleryDetail?.language.codes ?? []
+                    )
+                } else {
+                    Logger.info("analyzeImageForLiveText image not found", context: ["index": index])
+                }
+            case .failure(let error):
+                Logger.info("analyzeImageForLiveText failed", context: ["index": index, "error": error])
+            }
+        }
     }
 }
 
@@ -261,6 +302,8 @@ private struct HorizontalImageStack: View {
     private let imageURLs: [Int: URL]
     private let originalImageURLs: [Int: URL]
     private let loadingStates: [Int: LoadingState]
+    private let enablesLiveText: Bool
+    private let liveTextGroups: [Int: [LiveTextGroup]]
     private let fetchAction: (Int) -> Void
     private let refetchAction: (Int) -> Void
     private let prefetchAction: (Int) -> Void
@@ -270,17 +313,16 @@ private struct HorizontalImageStack: View {
     private let copyImageAction: (URL) -> Void
     private let saveImageAction: (URL) -> Void
     private let shareImageAction: (URL) -> Void
-    private let recognizeTextFromImageAction: (URL) -> Void
 
     init(
         index: Int, isDualPage: Bool, isDatabaseLoading: Bool, backgroundColor: Color,
         config: ImageStackConfig, imageURLs: [Int: URL], originalImageURLs: [Int: URL],
-        loadingStates: [Int: LoadingState], fetchAction: @escaping (Int) -> Void,
+        loadingStates: [Int: LoadingState], enablesLiveText: Bool,
+        liveTextGroups: [Int: [LiveTextGroup]], fetchAction: @escaping (Int) -> Void,
         refetchAction: @escaping (Int) -> Void, prefetchAction: @escaping (Int) -> Void,
         loadRetryAction: @escaping (Int) -> Void, loadSucceededAction: @escaping (Int) -> Void,
         loadFailedAction: @escaping (Int) -> Void, copyImageAction: @escaping (URL) -> Void,
-        saveImageAction: @escaping (URL) -> Void, shareImageAction: @escaping (URL) -> Void,
-        recognizeTextFromImageAction: @escaping (URL) -> Void
+        saveImageAction: @escaping (URL) -> Void, shareImageAction: @escaping (URL) -> Void
     ) {
         self.index = index
         self.isDualPage = isDualPage
@@ -290,6 +332,8 @@ private struct HorizontalImageStack: View {
         self.imageURLs = imageURLs
         self.originalImageURLs = originalImageURLs
         self.loadingStates = loadingStates
+        self.enablesLiveText = enablesLiveText
+        self.liveTextGroups = liveTextGroups
         self.fetchAction = fetchAction
         self.refetchAction = refetchAction
         self.prefetchAction = prefetchAction
@@ -299,7 +343,6 @@ private struct HorizontalImageStack: View {
         self.copyImageAction = copyImageAction
         self.saveImageAction = saveImageAction
         self.shareImageAction = shareImageAction
-        self.recognizeTextFromImageAction = recognizeTextFromImageAction
     }
 
     var body: some View {
@@ -320,6 +363,8 @@ private struct HorizontalImageStack: View {
             loadingState: loadingStates[index] ?? .idle,
             isDualPage: isDualPage,
             backgroundColor: backgroundColor,
+            enablesLiveText: enablesLiveText,
+            liveTextGroups: liveTextGroups[index] ?? [],
             refetchAction: refetchAction,
             loadRetryAction: loadRetryAction,
             loadSucceededAction: loadSucceededAction,
@@ -367,11 +412,6 @@ private struct HorizontalImageStack: View {
             } label: {
                 Label(R.string.localizable.readingViewContextMenuButtonShare(), systemSymbol: .squareAndArrowUp)
             }
-            Button {
-                recognizeTextFromImageAction(imageURL)
-            } label: {
-                Label("Text recognition", systemSymbol: .aMagnify)
-            }
         }
     }
 }
@@ -390,6 +430,8 @@ private struct ImageContainer: View {
     private let loadingState: LoadingState
     private let isDualPage: Bool
     private let backgroundColor: Color
+    private let enablesLiveText: Bool
+    private let liveTextGroups: [LiveTextGroup]
     private let refetchAction: (Int) -> Void
     private let loadRetryAction: (Int) -> Void
     private let loadSucceededAction: (Int) -> Void
@@ -398,7 +440,10 @@ private struct ImageContainer: View {
     init(
         index: Int, imageURL: URL?,
         loadingState: LoadingState,
-        isDualPage: Bool, backgroundColor: Color,
+        isDualPage: Bool,
+        backgroundColor: Color,
+        enablesLiveText: Bool,
+        liveTextGroups: [LiveTextGroup],
         refetchAction: @escaping (Int) -> Void,
         loadRetryAction: @escaping (Int) -> Void,
         loadSucceededAction: @escaping (Int) -> Void,
@@ -409,6 +454,8 @@ private struct ImageContainer: View {
         self.loadingState = loadingState
         self.isDualPage = isDualPage
         self.backgroundColor = backgroundColor
+        self.enablesLiveText = enablesLiveText
+        self.liveTextGroups = liveTextGroups
         self.refetchAction = refetchAction
         self.loadRetryAction = loadRetryAction
         self.loadSucceededAction = loadSucceededAction
@@ -437,7 +484,11 @@ private struct ImageContainer: View {
 
     var body: some View {
         if loadingState == .idle {
-            image(url: imageURL).scaledToFit()
+            image(url: imageURL).scaledToFit().overlay(
+                LiveTextView(liveTextGroups: liveTextGroups)
+                    .opacity(enablesLiveText ? 1 : 0)
+            )
+            .animation(.default, value: enablesLiveText)
         } else {
             ZStack {
                 backgroundColor
